requirements.txt
Flask==2.3.2
pymongo==4.7.2
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.4
